Created a working skeleton of code to scrape airbnbs Im feeling lucky section
Code is entirely procedural and not encapsulated yet, this is fine as this is the next thing to be done. Furthermore,
there is a slight issue with AirBNB blocking the scraper because of "Too Many Requests". This could be handled either
by dropping usage of BS4, using rotating Proxies (ideal), or implementing time gaps periodically qhilst scraping (not ideal).