Successfully encapsulated script into a class. Added comments and bits of docstrings.
As discussed on slack, the module has been encapsulated and organised nicely, with lots of commenting so now this is
super easy to read and understand. Now, it needs proper industrial strandard docstrings everywhere, and that is
very much a baseline functional product. But it would be fantastic if a rotating proxy could now be implemented
I am gong to have a crack at this over the coming week, whilst the essential jobs are being done.